LayoutTests/http/wpt/credential-management/credentialscontainer-get-basics.https.html

<!DOCTYPE html>
<title>Credential Management API: get() basics.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../webauthn/resources/util.js"></script>
<script>
promise_test(function(t) {
    return promise_rejects_dom(t, "NotSupportedError",
            navigator.credentials.get());
}, "navigator.credentials.get() with no argument.");

promise_test(function(t) {
    return promise_rejects_dom(t, "NotSupportedError",
            navigator.credentials.get({ }));
}, "navigator.credentials.get() with empty argument.");

promise_test(function(t) {
    return promise_rejects_js(t, TypeError,
            navigator.credentials.get({publicKey: "bogus publicKey data"}));
}, "navigator.credentials.get() with bogus publicKey data");

promise_test(function(t) {
    return promise_rejects_dom(t, "NotSupportedError",
            navigator.credentials.get({bogus_key: "bogus data"}));
}, "navigator.credentials.get() with bogus data");

promise_test(function(t) {
    var controller = new AbortController();
    var options = { signal: controller.signal };

    controller.abort();
    return promise_rejects_dom(t, "AbortError",
        navigator.credentials.get(options));
}, "navigator.credentials.get() with abort signal set");

promise_test(async (t) => {
    const options = {
        digital: { providers: [] },
        publicKey: {
            challenge: Base64URL.parse("MTIzNDU2"),
            timeout: 100
        },
    };
    const p = navigator.credentials.get(options);
    await promise_rejects_dom(t, "NotSupportedError", p);
    try {
        await p;
    } catch (e) {
        assert_equals(e.message, "Only one request type is supported at a time.");
    }
}, "navigator.credentials.get() with multiple request types.");

promise_test(async (t) => {
    const p =navigator.credentials.get({ mediation: "silent" });
    await promise_rejects_dom(t, "NotSupportedError", p);
    try {
        await p;
    } catch (e) {
        assert_equals(e.message, "Missing request type.");
    }
}, "navigator.credentials.get() with a single mediation argument.");

promise_test(async (t) => {
    const {signal} = new AbortController();
    const options = { signal, mediation: "silent" };
    const p = navigator.credentials.get(options);
    await promise_rejects_dom(t, "NotSupportedError", p);
    try {
        await p;
    } catch (e) {
        assert_equals(e.message, "Missing request type.");
    }
}, "navigator.credentials.get() with only a mediation and signal.");
</script>
